tests

+++ src/modules/user/services/delete-user.service.ts
@@ -2,7 +2,7 @@ import { Injectable } from '@nestjs/common';
 import { InjectModel } from '@nestjs/mongoose';
 import { Model } from 'mongoose';
 import { AppException, ErrorCode } from '../../../app.exception';
-import { DescendantLocationsFinderService } from '../../location/descendant-locations-finder.service';
+import { DescendantLocationCheckerService } from '../../location/descendant-location-checker.service';
 import { User } from '../user.schema';
 import { CurrentUser } from '../user.types';
 
@@ -10,18 +10,12 @@ import { CurrentUser } from '../user.types';
 export class DeleteUserService {
   constructor(
     @InjectModel(User.name) private readonly userModel: Model<User>,
-    private readonly descendantLocationsFinderService: DescendantLocationsFinderService,
+    private readonly descendantLocationCheckerService: DescendantLocationCheckerService,
   ) {}
 
-  async execute(
-    currentUser: CurrentUser,
-    id: string,
-  ): Promise<{ deleted: boolean }> {
+  async execute(currentUser: CurrentUser, id: string): Promise<{ deleted: boolean }> {
     if (currentUser.id === id) {
-      throw new AppException(
-        ErrorCode.USER_HAS_NO_ACCESS,
-        'You cannot delete your own account.',
-      );
+      throw new AppException(ErrorCode.USER_HAS_NO_ACCESS, 'You cannot delete your own account.');
     }
 
     const userExists = await this.userModel.findById(id).exec();
@@ -29,29 +23,15 @@ export class DeleteUserService {
       throw new AppException(ErrorCode.USER_NOT_FOUND, 'User does not exist.');
     }
 
-    await this.checkUserLocation(
+    const inLocationHierarchy = await this.descendantLocationCheckerService.execute(
       currentUser.locationId,
       userExists.locationId.toString(),
     );
+    if (!inLocationHierarchy) {
+      throw new AppException(ErrorCode.USER_HAS_NO_ACCESS, 'You cannot delete this user. This user is not in your location hierarchy.');
+    }
 
     await userExists.deleteOne().lean();
     return { deleted: true };
-  }
-
-  private async checkUserLocation(
-    currentUserLocationId: string,
-    locationId: string,
-  ) {
-    const descendantLocations =
-      await this.descendantLocationsFinderService.execute(
-        currentUserLocationId,
-      );
-
-    if (!descendantLocations.includes(locationId)) {
-      throw new AppException(
-        ErrorCode.USER_HAS_NO_ACCESS,
-        'You cannot delete this user. This user is not in your location hierarchy.',
-      );
-    }
   }
 }
